### 2025-01-05 #7629 @editorialbot
[PRE REVIEW]: MPI-Rockstar: a Hybrid MPI and OpenMP Parallel Implementation of the Rockstar Halo finder
**Submitting author:** @Tomoaki-Ishiyama (<a href="http://orcid.org/0000-0002-5316-9171">Tomoaki Ishiyama</a>)
**Repository:** https://github.com/Tomoaki-Ishiyama/mpi-rockstar
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
