### 2025-01-05 #7629 @warrickball
Hi @Tomoaki-Ishiyama, and thanks for your submission to JOSS. The nature of this submission looks a bit unusual so I'd like to clear up a few things before we consider continuing to review. Specifically, is the idea that JOSS reviews the submitted repo *and* the [original Rockstar code](https://bitbucket.org/gfcstanford/rockstar/)? Or is this submission in effect a standalone fork that extends the previous code?

The latter is in principle in scope for JOSS (see [*Should I write my own software or contribute to an existing package?*](https://joss.readthedocs.io/en/latest/submitting.html#should-i-write-my-own-software-or-contribute-to-an-existing-package)) but I notice that the new documentation is just the top-level README, which has some technical details but little context, and there don't appear to be any examples ready to run out-of-the-box.
