### 2025-01-10 #7629 @Tomoaki-Ishiyama
Hi @warrickball and thank you so much for your comments. This submission is in effect a standalone fork that extends the original Rockstar code.

Indeed, the current README contains little context. We will attempt to fix it as soon as possible.
