### 2025-01-15 #7629 @warrickball
Thanks. In that case, I'm going to flag this article for an editorial review of whether the extension to MPI meets JOSS requirements for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This should take about two weeks. Sorry for the extra wait!
