### 2025-01-16 #7629 @Tomoaki-Ishiyama
We believe that this code meets JOSS requirements because the extension to MPI and massively parallel environment took nearly a year and our community was eager for this extension. Thank you for your consideration.
