### 2025-01-31 #7629 @warrickball
Hi again @Tomoaki-Ishiyama, I'm pleased to say that the editorial board has agreed that the MPI parallelisation alone is a substantial contribution, so we'll continue our process. All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!
