### 2025-04-28 #7629 @Tomoaki-Ishiyama
@warrickball We have been waiting the editor assignment for more than 3 months. How much longer do we have to wait? Editors seem to be assigned for the most of papers submitted after us. Thank you for your assistance.
