### 2025-04-28 #7629 @warrickball
Hi @Tomoaki-Ishiyama, I do apologise for the very long wait. The Astronomy, Astrophysics & Space Science (AASS) track of JOSS has a substantial backlog, which is typically leading to a ~3 month wait from submission to an editor becoming available. MPI-Rockstar is now in effect at the front of the queue but we still need an editor with appropriate expertise to become available, which is why some more recent submissions (though many also >2 months old) have been picked up first.
