### 2025-05-18 #7629 @junghans
> [@junghans](https://github.com/junghans) I have just updated the draft following your comments. Please review it again.

Thanks for adding that section, could you please also add a figure showing the hybrid parallelization scheme?
