### 2025-05-20 #7629 @Tomoaki-Ishiyama
@junghans Thank you for your suggestion. I have just added a figure showing the hybrid parallelization scheme. Please review it again.
