### 2025-05-22 #7629 @phil-mansfield
Unfortunately, I won't have time to review this paper right now. (Which is a shame: it's cool!) Thank you for the invitation, though.
