### 2025-05-27 #8295 @editorialbot
[REVIEW]: MPI-Rockstar: a Hybrid MPI and OpenMP Parallel Implementation of the Rockstar Halo finder
**Submitting author:** @Tomoaki-Ishiyama (<a href="http://orcid.org/0000-0002-5316-9171">Tomoaki Ishiyama</a>)
**Repository:** https://github.com/Tomoaki-Ishiyama/mpi-rockstar
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.0
