### 2025-06-04 Tomoaki-Ishiyama/mpi-rockstar#10 @yymao
[JOSS review] Improve compiling instructions/procedure
> This issue is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/8295

- [ ] Is there any specific requirement on the MPI distributions (e.g., openmpi, mpich, intel-oneapi-mpi)? If so, please specify.
- [ ] The instruction should remind the user to set the compiler that is being used in `Makefile`.
- [ ] When `HDF5_INCLUDE` and `HDF5_LIB` are not set properly, Makefile should return an error, if possible.
- [ ] The instruction should mention that, if compiling is not successfully due to wrong paths/setting, once the issue is fixed in `Makefile`, the user should first run `make clean` before proceeding to recompile again.
