### 2025-06-04 Tomoaki-Ishiyama/mpi-rockstar#11 @yymao
[JOSS review] Include community guidelines
> This issue is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/8295

According to JOSS review checklist, in the documentation (README) there should be clear guidelines for third parties wishing to (1) Contribute to the software (2) Report issues or problems with the software (3) Seek support.

Please add these community guidelines.
