### 2025-06-06 Tomoaki-Ishiyama/mpi-rockstar#12 @yymao
[JOSS review] Comments on the associated paper
> This issue is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/8295

I have several comments regarding the paper for the authors to consider:

- [ ] In the Statement of Need section, it is mentioned that “... they have not been uniformly tested yet on the large-scale hybrid shared and distributed memory environments of modern supercomputers. MPI-Rockstar addresses these issues …” However, MPI-Rockstar does *not* address the issue that many halo finders have not been uniformly tested in large-scale HPC settings. That is definitely a task beyond the scope of this work.

- [ ] Related to above, the Statement of Need section should really mention why a MPI implementation is needed, given that Rockstar can be run on distributed machines already. This is somewhat addressed in the Parallelization section (around Lines 50-52). I’d suggest moving that discussion to Statement of Need, or forward reference it in the Statement of Need section.

- [ ] Still related to above, regarding the limitation of Rockstar, can you state more specifically in what scenario (e.g., beyond how many nodes, or for how many number of particles) that Rockstar’s socket implementation won’t perform well? This is an important argument for the statement of need. In other words, in what scenario(s) should people use MPI-Rockstar instead of Rockstar? In fact, in the Summary section (around Lines 26-27) there’s a mention of these numbers (# particles, # processes), but it wasn’t clear to me if those number are just examples, or if they are indeed the limitation thresholds of original Rockstar.

- [ ] Around Lines 73-74, it is mentioned that “MPI-Rockstar could run up to three times faster than the original Rockstar when compared in the same execution environment.” What is the setup (# particles, # processes) that this statement was tested on? I imagine the speed up (3x) is not the same for all possible setups, so it’s important to specify.

- [ ] The authors must have verified that the MPI-Rockstar gives the identical halo finding results as Rockstar (for a case where Rockstar can still manage). Please state so in the paper, including the verification test setup.

- [ ] Figure 1 caption: typo in “strategy”

- [ ] Figure 2: please explain the dashed lines.

- [ ] Figure 2: please use different line styles for the two boxes so that the plot is friendly to people with color vision deficiency.

- [ ] Figure 2: The $y$-axis the execution time for a single snapshot (at z=2), right? Maybe a better caption to remind the readers.

- [ ] Figure 2: Is it possible to put Rockstar on this plot (for the 2560^3 box, or even a smaller box say 1024^3)? (This is related to my question about the performance enhancement of MPI-Rockstar when compared to Rockstar.)

- [ ] Lines 67, 71, and 72: it seems that <sup> is not working in the PDF rendering. Maybe use $2560^3$ instead.
