### 2025-06-23 Tomoaki-Ishiyama/mpi-rockstar#13 @kyleaoman
[JOSS review] Compiler warnings
This issue doesn't block my JOSS review. Just wanted to flag a couple of warnings issue on compilation:

```
mpi_main.cpp: In function ‘void align_domain_particles(int, const float (*)[3], int64_t*, int64_t, int64_t, int64_t, const int*, float (*)[6])’:
mpi_main.cpp:404:16: warning: lambda capture initializers only available with ‘-std=c++14’ or ‘-std=gnu++14’ [-Wc++14-extensions]
  404 |               [all_samples = all_samples, axis = axis](int64_t a, int64_t b) {
      |                ^~~~~~~~~~~
mpi_main.cpp:404:43: warning: lambda capture initializers only available with ‘-std=c++14’ or ‘-std=gnu++14’ [-Wc++14-extensions]
  404 |               [all_samples = all_samples, axis = axis](int64_t a, int64_t b) {
      |                                           ^~~~
```
