### 2025-06-23 Tomoaki-Ishiyama/mpi-rockstar#10 @kyleaoman
- [x] The list of dependencies should be clearly stated. For example, on attempting to compile I get:
```
In file included from io/meta_io.c:22:
io/io_tipsy.h:4:10: fatal error: rpc/types.h: No such file or directory
    4 | #include <rpc/types.h>
      |          ^~~~~~~~~~~~~
compilation terminated.
```
- [x] Supported versions for MPI, HDF5 and any other relevant dependencies should be given.
- [x] It should not be necessary for the user to edit the Makefile directly (e.g. to choose a compiler, to remove `-I/usr/include/tirpc` as suggested in the instructions, or to specify the location of hdf5 libraries). Please move such configuration options to a configuration script, command line option, or similar.
