### 2025-06-23 Tomoaki-Ishiyama/mpi-rockstar#14 @kyleaoman
[JOSS review] Relationship to original rockstar
- [ ] Please state the relationship to the original rockstar implementation in the statement of need. Is mpi-rockstar intended to replace the socket-based implementation? Will both implementations receive ongoing maintenance?
- [ ] How will changes be kept in sync, where relevant?
- [ ] Crucially, the mpi-rockstar documentation currently depends on the rockstar documentation. If the two implementations start to drift apart this could cause significant confusion as options documented on rockstar might not be available on mpi-rockstar and there is no way to automatically monitor such discrepancies in the documentation. I leave it to the authors to decide on a solution, but one approach would be to fully document mpi-rockstar independently of the original documentation.
