### 2025-06-23 Tomoaki-Ishiyama/mpi-rockstar#15 @kyleaoman
[JOSS review] Functionality documentation
- [ ] Please state how to specify compile-time options (such as `OUTPUT_RVMAX`), giving an example. (It should not be necessary to edit the Makefile for this.)
- [ ] The option for the minimum number of particles for fitting the NFW profile should be moved from the source code to the config file. If there are other such hard-coded numbers that should be configurable please treat them similarly.
- [ ] In the readme, there is an asterix following `Note that each sub-directory is made automatically.*` in the description of `FILES_PER_SUBDIR_OUTPUT and SUBDIR_DIGITS_OUTPUT`. Is there meant to be a footnote? Or is this perhaps a typo in applying bold formatting?
