### 2025-06-23 Tomoaki-Ishiyama/mpi-rockstar#16 @kyleaoman
[JOSS review] Creation of directories
- [x] The description of the `OUTPUT_SUBDIR` option mentions that the subdirectories must be created before running the code. It would be much better to avoid this constraint - users on HPC systems will normally queue a job that might take several days to begin execution. A crash due to forgetting to create directories is a significant but avoidable setback. If it is not feasible to create any missing directories during execution, consider providing a dry-run mode that can check for presence of needed files & directories with minimal resources (e.g. so that users would not normally need to submit it to a job queue on an HPC system).
