### 2025-06-23 Tomoaki-Ishiyama/mpi-rockstar#17 @kyleaoman
[JOSS review] Examples
- [x] The sample file name `256_50hMpc_z0.kyf` does not match the file that the code looks for (`256_50hMpc_050.kyf`).
- [x] Giving approximate system requirements to run the examples would also be useful (more generally X particles = Y GB memory footprint is a helpful rule of thumb to have in the docs). I now just happen to know that the 1024^3 example just *barely* fits in 64GB memory (sometimes doesn't...possibly depending on "random" order of calls to MPI?) and runs in ~400s (per snapshot) on 16 ranks (32 cores). There may have been a bit of swapping (8 GB swap space), I wasn't watching this closely.
