### 2025-06-23 Tomoaki-Ishiyama/mpi-rockstar#18 @kyleaoman
[JOSS review] Unpredictable memory usage
Trying a few runs of the larger example, I get occasional crashes due to running out of memory (on my 64GB machine). On one run for instance I get:

```
txwx36@iccpc2> mpiexec -n 16 mpi-rockstar -c parallel_1024.cfg
#The_number_of_threads= 4
#OUTPUT_RVMAX on
#OUTPUT_INERTIA_TENSOR on
#OUTPUT_NFW_CHI2 on
[     0s] Start processing snapshot 49 with 16 procs.
[    17s] 16 procs read 640 blocks for snapshot 49.
#chunks=(4,2,2)
[    27s] Transferring0...
[    29s] Transferring particles to writers...
[    41s] Analyzing for FoF groups...
[   117s] Transferring boundary particles between writers...
[   118s] Linking boundary particles...
[   118s] Analyzing for halos / subhalos...
[   408s] Output halos...
[   411s] [Success] Done with snapshot 49.
[   411s] Start processing snapshot 50 with  16 procs.
[   431s] 16 procs read 640 blocks for snapshot 50.
#chunks=(4,2,2)
[   441s] Transferring0...
--------------------------------------------------------------------------
prterun noticed that process rank 0 with PID 2389672 on node iccpc2 exited on
signal 9 (Killed).
--------------------------------------------------------------------------
```

While removing the `Out` directory and running again gave a clean run through both snapshots. It seemed like the peak memory usage (and therefore the risk of crashing) was between the `Transferring0...` and `Transferring particles to writers...` messages. If the memory varies by perhaps a few percent then this is not a big deal, but I think that it may be more than this since I have 8GB swap available so it should have slowed down significantly and swapped if a little more memory was needed, the immediate crash is suggestive of a new large allocation (>8 GB) before existing allocations are cleaned up.

Another crashed run came with some additional error messages:
```
txwx36@iccpc2> mpiexec -n 16 mpi-rockstar -c parallel_1024.cfg
#The_number_of_threads= 4
#OUTPUT_RVMAX on
#OUTPUT_INERTIA_TENSOR on
#OUTPUT_NFW_CHI2 on
[     0s] Start processing snapshot 49 with 16 procs.
[    18s] 16 procs read 640 blocks for snapshot 49.
#chunks=(4,2,2)
[    31s] Transferring0...
[1750712326.049228239] iccpc2.phyast.dur.ac.uk:rank2.mpi-rockstar: Reading from remote process' memory failed. Disabling CMA support
libfabric:2388550:1750712326::psm3:av:psmx3_epid_to_epaddr():234<warn> iccpc2.phyast.dur.ac.uk:rank8: psm3_ep_connect returned error Operation timed out, remote epid=0x81eab8c70000017c:da14c39f:0.Try setting FI_PSM3_CONN_TIMEOUT to a larger value (current: 10 seconds). Aborting
[iccpc2:2388550] *** Process received signal ***
[iccpc2:2388550] Signal: Aborted (6)
[iccpc2:2388550] Signal code:  (-6)
[iccpc2:2388550] [ 0] /lib64/libc.so.6(+0x1a050) [0x7f0fef541050]
[iccpc2:2388550] [ 1] /lib64/libc.so.6(+0x72f54) [0x7f0fef599f54]
[iccpc2:2388550] [ 2] /lib64/libc.so.6(gsignal+0x1e) [0x7f0fef540f9e]
[iccpc2:2388550] [ 3] /lib64/libc.so.6(abort+0xdf) [0x7f0fef528942]
[iccpc2:2388550] [ 4] /lib64/libfabric.so.1(+0x34351d) [0x7f0feef4351d]
[iccpc2:2388550] [ 5] /lib64/libfabric.so.1(+0x328f03) [0x7f0feef28f03]
[iccpc2:2388550] [ 6] /lib64/libfabric.so.1(+0x34c40a) [0x7f0feef4c40a]
[iccpc2:2388550] [ 7] /usr/lib64/openmpi/lib/libmpi.so.40(+0x19173f) [0x7f0fef99173f]
[iccpc2:2388550] [ 8] /usr/lib64/openmpi/lib/libmpi.so.40(mca_pml_cm_start+0x183) [0x7f0fefa19763]
[iccpc2:2388550] [ 9] /usr/lib64/openmpi/lib/libmpi.so.40(ompi_coll_base_alltoallv_intra_basic_linear+0x2ad) [0x7f0fef8cdfcd]
[iccpc2:2388550] [10] /usr/lib64/openmpi/lib/libmpi.so.40(ompi_coll_tuned_alltoallv_intra_dec_fixed+0x42) [0x7f0fef8f3ea2]
[iccpc2:2388550] [11] /usr/lib64/openmpi/lib/libmpi.so.40(MPI_Alltoallv+0x66f) [0x7f0fef84feaf]
[iccpc2:2388550] [12] mpi-rockstar() [0x43de97]
[iccpc2:2388550] [13] mpi-rockstar() [0x43a1d6]
[iccpc2:2388550] [14] mpi-rockstar() [0x43da68]
[iccpc2:2388550] [15] mpi-rockstar() [0x401b47]
[iccpc2:2388550] [16] /lib64/libc.so.6(+0x3248) [0x7f0fef52a248]
[iccpc2:2388550] [17] /lib64/libc.so.6(__libc_start_main+0x8b) [0x7f0fef52a30b]
[iccpc2:2388550] [18] mpi-rockstar() [0x401bb5]
[iccpc2:2388550] *** End of error message ***
--------------------------------------------------------------------------
prterun noticed that process rank 1 with PID 2388543 on node iccpc2 exited on
signal 9 (Killed).
--------------------------------------------------------------------------
```

This is with:
```
txwx36@iccpc2> mpicc --version
gcc (GCC) 14.3.1 20250523 (Red Hat 14.3.1-1)
```
and:
```
txwx36@iccpc2> mpiexec --version
mpiexec (Open MPI) 5.0.5
```
