### 2025-06-26 #8295 @junghans
@Tomoaki-Ishiyama could you please have a look at those issue?
