### 2025-06-26 Tomoaki-Ishiyama/mpi-rockstar#17 @junghans
Related to https://github.com/openjournals/joss-reviews/issues/8295
