### 2025-07-05 cb980f5
Append .hdf5 suffix to GADGET-4 snapshot filename if ENABLE_HDF5
