+++ src/io/meta_io.c
@@ -149,6 +149,7 @@ void read_particles(char *filename) {
         load_particles(filename, &p, &num_p);
     else if (!strncasecmp(FILE_FORMAT, "GADGET4", 7)) {
 #ifdef ENABLE_HDF5
+		strstr(filename, ".hdf5") || strcat(filename, ".hdf5");
         load_particles_gadget4(filename, &p, &num_p);
         gadget = 1;
 #else
