### 2025-07-05 8b72a7e
Add io_gadget2_hdf5, modify meta_io, config.template and update Makefile
