+++ src/config.template.h
@@ -42,6 +42,12 @@ integer(GADGET4_NTYPES, 6);
 real(GADGET4_MASS_CONVERSION, 1e10);
 real(GADGET4_LENGTH_CONVERSION, 1.0);
 integer(GADGET4_DM_PARTTYPE, 1);
+
+integer(GADGET2_ID_BYTES, 8);
+integer(GADGET2_NTYPES, 6);
+real(GADGET2_MASS_CONVERSION, 1e10);
+real(GADGET2_LENGTH_CONVERSION, 1.0);
+integer(GADGET2_DM_PARTTYPE, 1);
 #endif /* ENABLE_HDF5 */
 
 real(TIPSY_LENGTH_CONVERSION, 1.0);
