+++ src/io/io_gadget2_hdf5.c
@@ -1,0 +1,262 @@
+/*
+ * Gadget-2 HDF5 I/O for Rockstar
+ * Minh Nguyen (nhat.minh.nguyen@ipmu.jp)
+ */
+#ifdef ENABLE_HDF5
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <math.h>
+#include <stdint.h>
+#include <assert.h>
+#include <string.h>
+#include <hdf5.h> /* HDF5 required */
+#include "io_hdf5.h"
+#include "io_gadget2_hdf5.h"
+#include "io_util.h"
+#include "../universal_constants.h"
+#include "../check_syscalls.h"
+#include "../config_vars.h"
+#include "../config.h"
+#include "../particle.h"
+
+void gadget2_readdataset_float(hid_t HDF_FileID, char *filename, char *gid,
+                               char *dataid, struct particle *p, int64_t to_read,
+                               int64_t offset, int64_t stride) {
+    int64_t   width   = H5Tget_size(H5T_NATIVE_FLOAT);
+    void     *buffer  = check_malloc_s(buffer, to_read, width * stride);
+    float    *fbuffer = buffer;
+
+    hid_t HDF_GroupID     = check_H5Gopen(HDF_FileID, gid, filename);
+    hid_t HDF_DatasetID   = check_H5Dopen(HDF_GroupID, dataid, gid, filename);
+    hid_t HDF_DataspaceID = check_H5Dget_space(HDF_DatasetID);
+
+    check_H5Sselect_all(HDF_DataspaceID);
+    hssize_t npoints = H5Sget_select_npoints(HDF_DataspaceID);
+
+    if (npoints != to_read * stride) {
+        fprintf(stderr,
+                "[Error] dataspace %s/%s in HDF5 file %s not expected size!\n  "
+                "(Actual size = %" PRId64 " elements; expected size = %" PRId64
+                " elements\n",
+                gid, dataid, filename, (int64_t)(npoints), stride * to_read);
+        exit(1);
+    }
+
+    check_H5Dread(HDF_DatasetID, H5T_NATIVE_FLOAT, buffer, dataid, gid, filename);
+
+    H5Sclose(HDF_DataspaceID);
+    H5Dclose(HDF_DatasetID);
+    H5Gclose(HDF_GroupID);
+
+    for (int64_t i = 0; i < to_read; i++)
+        memcpy(((char *)&(p[i])) + offset, fbuffer + (i * stride), stride * width);
+
+    free(buffer);
+}
+
+void gadget2_readdataset_ID_uint32(hid_t HDF_FileID, char *filename, char *gid,
+                                   char *dataid, struct particle *p, int64_t to_read,
+                                   int64_t offset, int64_t stride) {
+    int64_t   width   = H5Tget_size(H5T_NATIVE_UINT32);
+    void     *buffer  = check_malloc_s(buffer, to_read, width * stride);
+    uint32_t *ibuffer = buffer;
+
+    hid_t HDF_GroupID     = check_H5Gopen(HDF_FileID, gid, filename);
+    hid_t HDF_DatasetID   = check_H5Dopen(HDF_GroupID, dataid, gid, filename);
+    hid_t HDF_DataspaceID = check_H5Dget_space(HDF_DatasetID);
+
+    check_H5Sselect_all(HDF_DataspaceID);
+    hssize_t npoints = H5Sget_select_npoints(HDF_DataspaceID);
+
+    if (npoints != to_read * stride) {
+        fprintf(stderr,
[…]
