### 2025-07-06 48577a6
Ensure TOTAL_PARTICLES and PARTICLE_MASS are correctly derived from header
