### 2025-07-06 cea24b3
Ensure PARTICLE_MASS derived from massTable in header
