### 2025-07-07 4e389fc
Remove check and force .hdf5 suffix at the end of filename
