+++ src/io/meta_io.c
@@ -150,7 +150,6 @@ void read_particles(char *filename) {
         load_particles(filename, &p, &num_p);
     else if (!strncasecmp(FILE_FORMAT, "GADGET4", 7)) {
 #ifdef ENABLE_HDF5
-		strstr(filename, ".hdf5") || strcat(filename, ".hdf5");
         load_particles_gadget4(filename, &p, &num_p);
         gadget = 1;
 #else
@@ -160,7 +159,6 @@ void read_particles(char *filename) {
 #endif
 	} else if (!strncasecmp(FILE_FORMAT, "GADGET2_HDF5", 12)) {
 #ifdef ENABLE_HDF5
-		strstr(filename, ".hdf5") || strcat(filename, ".hdf5");
         load_particles_gadget2_hdf5(filename, &p, &num_p);
         gadget = 1;
 #else
