### 2025-07-07 5fc53d8
Remove extra H5Gclose(HDF_Header)
