### 2025-08-31 Tomoaki-Ishiyama/mpi-rockstar#16 @stanaka2
Thanks for reporting. This is now fixed in commit 5fee042.
