### 2025-09-28 7ed0934
Fix compile of find_parents
