### 2026-03-13 Tomoaki-Ishiyama/mpi-rockstar#10 @Tomoaki-Ishiyama
We have changed the style of documentation from README.md to readthedocs. Following the suggestion in #14, we have also reconstructed the documentation as it is independent of the original documentation.

In the new documentation, we have specified those things raised by referees. We have also  provided both Makefile and a configure script. When using the configure script, it returns an error when HDF5_INCLUDE and HDF5_LIB are not set properly.
