### 2026-03-13 Tomoaki-Ishiyama/mpi-rockstar#11 @Tomoaki-Ishiyama
We have added guidelines in the Policy section of the documentation.
