### 2026-03-13 Tomoaki-Ishiyama/mpi-rockstar#12 @Tomoaki-Ishiyama
> In the Statement of Need section, it is mentioned that “... they have not been uniformly tested yet on the large-scale hybrid shared and distributed memory environments of modern supercomputers. MPI-Rockstar addresses these issues …” However, MPI-Rockstar does not address the issue that many halo finders have not been uniformly tested in large-scale HPC settings. That is definitely a task beyond the scope of this work.

MPI-Rockstar demonstrates scaling in these environments. We confirm that MPI-Rockstar can analyze 2 trillion particle simulations on 16,384 nodes (786,432 CPU cores) of Fugaku. For this analysis, we used 32,768 MPI processes and 24 OpenMP threads per process. Other finders have not been uniformly tested yet on such large-scale hybrid shared and distributed memory environments.

> Still related to above, regarding the limitation of Rockstar, can you state more specifically in what scenario (e.g., beyond how many nodes, or for how many number of particles) that Rockstar’s socket implementation won’t perform well? This is an important argument for the statement of need. In other words, in what scenario(s) should people use MPI-Rockstar instead of Rockstar? In fact, in the Summary section (around Lines 26-27) there’s a mention of these numbers (# particles, # processes), but it wasn’t clear to me if those number are just examples, or if they are indeed the limitation thresholds of original Rockstar.

The main scaling bottleneck in Rockstar is that, even when the number of sockets is unlimited, a single-threaded server process has to coordinate between all the worker processes.  This starts to become a bottleneck around 10,000 processors. Besides, the number of sockets (file descriptors) can be issued simultaneously strongly depends on the operation system and settings of a supercomputer where Rockstar is executed. We have added these explanations in the text.

> The authors must have verified that the MPI-Rockstar gives the identical halo finding results as Rockstar (for a case where Rockstar can still manage). Please state so in the paper, including the verification test setup.

For the test of 1024^3 particles in a 250 Mpc/h at z=2, the difference of halo mass function is below 0.1% except for the massive and less massive end, where the statistics are infulenced by the small halo counts and resolution, respectively. Please note that Rockstar is not deterministic, and so identical halo catalogs are generally not possible to obtain. We have added this description in the paper.

> Figure 2: Is it possible to put Rockstar on this plot (for the 2560^3 box, or even a smaller box say 1024^3)? (This is related to my question about the performance enhancement of MPI-Rockstar when compared to Rockstar.)

Unfortunately, the original Rockstar could not run on the supercomputer that used to measure the performance of this plot. Instead, we compare both in a workstation, and specify the detail.

We have also modified the paper following all of other comments in this thread.
