### 2026-03-13 Tomoaki-Ishiyama/mpi-rockstar#14 @Tomoaki-Ishiyama
MPI-Rockstar is not intended to replace the original implementation. All the changes are not kept in sync. Following the suggestion, we have reconstructed the documentation as it is independent of the original documentation. (Please also see the thread #10).
