### 2026-03-13 Tomoaki-Ishiyama/mpi-rockstar#15 @Tomoaki-Ishiyama
We have added the description of how to specify compile-time options in the new documentation.

We have moved  the options related to the NFW fitting to the config file. There are other hard-coded numbers. However, since these do not affect the output and are performance-related, we deliberately leave them as is rather than moving them to the config file.

Indeed, the asterisk was a typo. Fixed it.
