### 2026-03-13 Tomoaki-Ishiyama/mpi-rockstar#17 @Tomoaki-Ishiyama
We have changed the filename as it matches the file the code looks for.

In terms of memory usage, MPI-Rockstar will use about 60 bytes / particle maximum total for a cosmological simulation. In addition, other memory for halos and buffer regions is necessary. Load imbalance increases memory usage depending on the simulation box size and resolution. Therefore, we can write the minimum memory requirement only. We have added the description in the new documentation as “In terms of memory usage, MPI-Rockstar will use about 60 bytes / particle maximum total for a cosmological simulation. Thus, if you have a 1024^3 particle simulation and 2GB of memory available per processor, you should plan on using at least 32 CPUs in parallel. "
