### 2026-03-13 Tomoaki-Ishiyama/mpi-rockstar#18 @Tomoaki-Ishiyama
The reply to #17 is also a reply to this thread.
