### 2026-03-16 9bcc7b2
Fix formatting of note in README.md
