### 2026-03-16 #8295 @kyleaoman
> My review is now concluded with the exception of trying to reproduce one of the scaling runs. I've opened some issues: [Tomoaki-Ishiyama/mpi-rockstar#13](https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/13) [Tomoaki-Ishiyama/mpi-rockstar#14](https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/14) [Tomoaki-Ishiyama/mpi-rockstar#15](https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/15) [Tomoaki-Ishiyama/mpi-rockstar#16](https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/16) [Tomoaki-Ishiyama/mpi-rockstar#17](https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/17) [Tomoaki-Ishiyama/mpi-rockstar#18](https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/18) and added some items to issue: [Tomoaki-Ishiyama/mpi-rockstar#10](https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/10)
>
> To have a go at reproducing the scaling run, I would need a copy of the snapshot that was used, and the config would be helpful as well. Since the snapshot is large, transferring it through Globus would be preferred. The target system has a subscription so a Globus Connect Personal endpoint (can be user-installed) on the source system is sufficient if there is no subscription already at that end.

I've had a look at the changes, the issues that I'd flagged in my review are resolved to my satisfaction. I've closed the corresponding github issues (@Tomoaki-Ishiyama if you prefer to re-open any of them feel free to do so).

My offer to try reproducing one of the scaling runs is still available, but would need the data made available as mentioned above.
