### 2026-03-17 #8295 @Tomoaki-Ishiyama
@kyleaoman  I'm sorry for missing your message about Globus. I appreciate the suggestion regarding Globus. I tried to setup Globus on the server hosting the data. However, I found that the internal security policy prevents the deployment of a Globus endpoint.

Given that the data size per snapshot is about 30GB, I believe that https-based access remains the most feasible option under this  technical constraint. To ensure data integrity, I have provided MD5 checksums in the parent directory, allowing users to verfy their downloads.
