### 2026-03-17 #8295 @kyleaoman
Ok, I will give it a try on [cosma](https://cosma.readthedocs.io/en/latest/).
