### 2026-03-17 #8295 @kyleaoman
@Tomoaki-Ishiyama I've tried analysing the 1024^3 snapshot 49. I get:
```
[     0s] Start processing snapshot 49 with 8 procs.
[    24s] 8 procs read 640 blocks for snapshot 49.
[    38s] Transferring0...
[    40s] Transferring particles to writers...
[    59s] Analyzing for FoF groups...
[   192s] Transferring boundary particles between writers...
[   193s] Linking boundary particles...
[   193s] Analyzing for halos / subhalos...
[   517s] Output halos...
[   523s] [Success] Done with snapshot 49.
```
With:
```
#!/bin/bash -l
#SBATCH --ntasks 8
#SBATCH --tasks-per-node 4
#SBATCH --cpus-per-task 8
#SBATCH -J mpirockstar
#SBATCH -o %j.oo
#SBATCH -e %j.eo
#SBATCH -p cosma8
#SBATCH -A dp004
#SBATCH -t 00:20:00
#SBATCH --mail-user=kyle.a.oman@durham.ac.uk
#SBATCH --mail-type=END
#SBATCH --no-requeue

export OMP_NUM_THREADS=16
export LD_LIBRARY_PATHS=$LD_LIBRARY_PATHS:/cosma/home/durham/dc-oman1/lib

mpiexec -n $SLURM_NTASKS ./mpi-rockstar -c parallel_1024.cfg
```
If I understand correctly that should be 8 MPI ranks x 16 cores each = 128 cores. This system has 128 cores/node so I could have fit the entire job on 1 node but split it across two because otherwise MPI communication is artificially fast (because the node communicates with itself).

Seems a bit slow (paper said 200s with 64 cores at z=0) but I haven't been terribly careful with optimization (e.g. no additional flags set when compiling, and I could probably cut that read time down considerably). Had a quick try running classic rockstar on the same input with a copy I had compiled ages ago but it's crashed (probably linked against libraries that no longer exist). I'll fix that and should then be able to make a fair comparison.
