### 2026-03-18 #8295 @Tomoaki-Ishiyama
@kyleaoman Thank you so much for your reports. FYI, this is our result using 64 cores (32 MPI processes, 2 threads/process). For the optimization, I added "-O3 -fopenmp".

[     0s] Start processing snapshot 50 with 64 procs.
[    30s] 64 procs read 640 blocks for snapshot 50.
[    39s] Transferring0...
[    43s] Transferring particles to writers...
[    51s] Analyzing for FoF groups...
[    88s] Transferring boundary particles between writers...
[    88s] Linking boundary particles...
[    88s] Analyzing for halos / subhalos...
[   189s] Output halos...
[   221s] Loading merger tree information...
[   222s] Constructing merger tree...
[   223s] [Success] Done with snapshot 50.
[   223s] [Finished]
