### 2026-03-18 #8295 @Tomoaki-Ishiyama
Sorry, That is a result using 64 MPI processes (1 thread/process). The below is one using 32 MPI processes (2 threads/process).

[     0s] Start processing snapshot 50 with 32 procs.
[     4s] 32 procs read 640 blocks for snapshot 50.
[    11s] Transferring0...
[    14s] Transferring particles to writers...
[    24s] Analyzing for FoF groups...
[    77s] Transferring boundary particles between writers...
[    77s] Linking boundary particles...
[    78s] Analyzing for halos / subhalos...
[   173s] Output halos...
[   204s] Loading merger tree information...
[   204s] Constructing merger tree...
[   207s] [Success] Done with snapshot 50.
[   207s] [Finished]
