### 2026-03-18 #8295 @kyleaoman
Ok, with more MPI ranks it does indeed run much faster. This:
```
#!/bin/bash -l
#SBATCH --ntasks 64
#SBATCH --tasks-per-node 32
#SBATCH --cpus-per-task 1
#SBATCH -J mpirockstar
#SBATCH -o %j.oo
#SBATCH -e %j.eo
#SBATCH -p cosma8
#SBATCH -A dp004
#SBATCH -t 00:20:00
#SBATCH --mail-user=kyle.a.oman@durham.ac.uk
#SBATCH --mail-type=END
#SBATCH --no-requeue

export OMP_NUM_THREADS=1
export LD_LIBRARY_PATHS=$LD_LIBRARY_PATHS:/cosma/home/durham/dc-oman1/lib

mpiexec -n $SLURM_NTASKS ./mpi-rockstar -c parallel_1024.cfg
```
Leads to:
```
[     0s] Start processing snapshot 49 with 64 procs.
[     3s] 64 procs read 640 blocks for snapshot 49.
[     6s] Transferring0...
[     7s] Transferring particles to writers...
[    11s] Analyzing for FoF groups...
[    30s] Transferring boundary particles between writers...
[    30s] Linking boundary particles...
[    30s] Analyzing for halos / subhalos...
[    94s] Output halos...
[    95s] [Success] Done with snapshot 49.
```
This does illustrate a gap in the documentation: it would be good to add to the docs pages (and perhaps to the paper) some guidance on how to allocate MPI ranks & threads. For other codes I work with it is better to minimize the number of MPI ranks (ideally 1/node) and then parallelize at the node level with threads (e.g. 128/node). For MPIRockstar, however, it seems like one should use as many MPI ranks as possible and only use threads... when, actually? Is there a scenario where using half as many MPI ranks each using twice as many threads is optimal? This is the type of question that should be answered in the documentation.
