### 2026-03-18 Tomoaki-Ishiyama/mpi-rockstar#11 @yymao
Implemented in https://mpi-rockstar.readthedocs.io/en/latest/policy.html
