### 2026-03-18 Tomoaki-Ishiyama/mpi-rockstar#12 @yymao
Thanks for the updated draft @Tomoaki-Ishiyam. It addresses most of my comments; however, the issues related to the captions of the two figures were not addressed. Can you address those so that I can complete the review?
