### 2026-03-18 #8295 @yymao
I have reviewed the revised paper and the new documentation. Most of my comments have been addressed, except for the ones related to figure captions as listed in https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/12.
