### 2026-03-20 Tomoaki-Ishiyama/mpi-rockstar#12 @Tomoaki-Ishiyama
@yymao Thank you so much for your review. I forgot to commit a new Figure 2. Now it should work.
https://github.com/Tomoaki-Ishiyama/mpi-rockstar/blob/joss/paper/paper.md

> Figure 2: Is it possible to put Rockstar on this plot (for the 2560^3 box, or even a smaller box say 1024^3)? (This is related to my question about the performance enhancement of MPI-Rockstar when compared to Rockstar.)

Unfortunately, the original Rockstar could not run on the supercomputer that used to measure the performance of this plot. Instead, we compare both in a workstation, and specify the detail.
