### 2026-03-20 #8295 @Tomoaki-Ishiyama
@kyleaoman Thanks for your suggestion. We agree this is a documentation gap. MPI-Rockstar is a hybrid MPI/OpenMP code, but the current documentation does not provide practical guidance on choosing MPI ranks vs OpenMP threads. But this guidance is too technical, I have added it to only the document (not paper).
https://mpi-rockstar.readthedocs.io/en/latest/get_started.html#performance-tips
In principle, because the optimal MPI rank / OpenMP thread configuration depends strongly on the target system, it is difficult to provide the optimal choice quantitatively. Instead, I have added a general guide.

@yymao Thank you for your review. I have replied it in https://github.com/Tomoaki-Ishiyama/mpi-rockstar/issues/12.
