### 2026-03-24 #8295 @Tomoaki-Ishiyama
@junghans I have checked the authors' information and released the software as v1.0.0. DOI is "10.5281/zenodo.19199548 ".
