### 2026-03-24 #8295 @Tomoaki-Ishiyama
The paper pdf contained unnecessary boldface. I have removed them and re-generated the pdf.
