### 2026-03-27 #8295 @warrickball
Hi @Tomoaki-Ishiyama, I wanted to let you know that the next step is for me (as associate editor-in-chief for this track) to request or propose (via PR) any finishing touches on the paper before we publish, but I'm about to go on vacation for 3 weeks. I've started doing a few preliminary checks and we should be able to wrap this up quite promptly once I'm back. Sorry about the delay!

In the meantime, I noticed that many of the references in the paper are missing journal titles. This is probably because you've wisely gathered the BibTeX entries from NASA ADS but our pipeline doesn't recognise its journal name macros (e.g. `\apj` instead of *The Astrophysical Journal*). Could you fix these in the paper? You can run `@editorialbot regenerate pdf` to render the updates if you like but that's not necessary, since I'll render it repeatedly locally when I get back to this. Note that you should write the journal name out in full because JOSS is an interdisciplinary journal. (So use e.g. *The Astrophysical Journal* instead of *ApJ*.)

Also, @yymao, would you mind ticking off the remaining items in [your review checklist](https://github.com/openjournals/joss-reviews/issues/8295#issuecomment-2922738915) just to confirm that you've completed your review? Thanks!
